+++ lib/antigravity/tool.rb
@@ -2,20 +2,16 @@
 
 module Antigravity
   class Tool
+    include Emojifiable
+
     class << self
       def inherited(subclass)
         subclass.extend(ClassMethods)
       end
-
-      def emoji
-        Antigravity.emoji(:tool)
-      end
     end
 
     module ClassMethods
-      def emoji
-        Antigravity.emoji(:tool)
-      end
+      include Emojifiable::ClassMethods
 
       def name(val = nil, desc: nil)
         if val
@@ -67,10 +63,6 @@ module Antigravity
           }
         }
       end
-    end
-
-    def emoji
-      self.class.emoji
     end
 
     def tool_name